+++ ScenarioGUI/utils/load_config_file.py
@@ -15,8 +15,6 @@ def load(gui_file: str | Path):
         globs.get_path_for_file(Path(gui_file).parent.parent, config["DEFAULT"]["PATH_2_ICONS"]).joinpath(config["DEFAULT"]["PATH_2_ICONS"]), "icons"
     )
 
-    globs.VERSION = globs.find_version(globs.get_path_for_file(Path(gui_file).parent.parent, "setup.cfg"))
-
     globs.WHITE = config["COLORS"]["WHITE"]
     globs.LIGHT = config["COLORS"]["LIGHT"]
     globs.LIGHT_SELECT = config["COLORS"]["LIGHT_SELECT"]
@@ -31,3 +29,8 @@ def load(gui_file: str | Path):
     globs.FILE_EXTENSION = config["DEFAULT"]["FILE_EXTENSION"]
     globs.GUI_NAME = config["DEFAULT"]["GUI_NAME"]
     globs.ICON_NAME = config["DEFAULT"]["ICON_NAME"]
+    try:
+        globs.VERSION = config["DEFAULT"]["VERSION"]
+    except KeyError:
+        globs.VERSION = globs.find_version(globs.get_path_for_file(Path(gui_file).parent.parent, "setup.cfg"))
+