+++ tests/test_main_window_functions/test_global_settings.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from configparser import ConfigParser
 from pathlib import Path
 from platform import system
 
@@ -12,6 +13,8 @@ from ScenarioGUI import load_config
 
 from ..starting_closing_tests import close_tests, start_tests
 
+path = Path(__file__).parent.parent
+
 
 def test_global_settings(qtbot):  # noqa: PLR0915
     """
@@ -23,10 +26,13 @@ def test_global_settings(qtbot):  # noqa: PLR0915
         bot for the GUI
     """
     # init gui window
-    load_config(Path(".").absolute().joinpath("./tests/gui_config.ini") if Path(".").absolute().joinpath("./tests/gui_config.ini").exists() else Path(
-        "..").absolute().joinpath("./tests/gui_config.ini"))
+    load_config(
+        Path(".").absolute().joinpath("./tests/gui_config.ini")
+        if Path(".").absolute().joinpath("./tests/gui_config.ini").exists()
+        else Path("..").absolute().joinpath("./tests/gui_config.ini")
+    )
     main_window = start_tests(qtbot)
-    assert main_window.dia.font().family() == globs.FONT == ("Arial" if system() == "Windows" else 'Helvetica')
+    assert main_window.dia.font().family() == globs.FONT == ("Arial" if system() == "Windows" else "Helvetica")
     assert main_window.dia.font().pointSize() == (11 if system() == "Windows" else 14) == globs.FONT_SIZE
 
     assert globs.WHITE == "rgb(255, 255, 255)"
@@ -43,8 +49,10 @@ def test_global_settings(qtbot):  # noqa: PLR0915
     assert globs.ICON_NAME == "icon.svg"
     assert globs.path.joinpath(".").parent if "tests" in f"{Path('.').absolute()}" else globs.path.joinpath(".") == globs.FOLDER
     # test get_path_for_file function
-    assert globs.get_path_for_file(globs.path.joinpath("./ScenarioGUI/gui_classes/gui_structure_classes"), "gui_config.ini") in [globs.path.parent.joinpath(
-        "." if "tests" in f"{Path('.').absolute()}" else "./examples"), globs.path.parent.joinpath("." if "tests" in f"{Path('.').absolute()}" else "./tests")]
+    assert globs.get_path_for_file(globs.path.joinpath("./ScenarioGUI/gui_classes/gui_structure_classes"), "gui_config.ini") in [
+        globs.path.parent.joinpath("." if "tests" in f"{Path('.').absolute()}" else "./examples"),
+        globs.path.parent.joinpath("." if "tests" in f"{Path('.').absolute()}" else "./tests"),
+    ]
     # test combine window settings
     check_font(main_window.push_button_cancel, globs.FONT_SIZE, globs.FONT, True)
     check_font(main_window.push_button_start_single, globs.FONT_SIZE, globs.FONT, True)
@@ -83,13 +91,24 @@ def test_global_settings(qtbot):  # noqa: PLR0915
     check_font(main_window.gui_structure.aim_add.widget, globs.FONT_SIZE, globs.FONT, True)
     check_font(main_window.gui_structure.category_inputs.label, globs.FONT_SIZE, globs.FONT, True)
     check_font(main_window.gui_structure.flex_option.label, globs.FONT_SIZE, globs.FONT, True)
-    [check_font(widget, globs.FONT_SIZE, globs.FONT, isinstance(widget, QtW.QPushButton)) for widget in main_window.gui_structure.flex_option.frame.children()
-     if
-     isinstance(widget, QtW.QWidget)]
+    [
+        check_font(widget, globs.FONT_SIZE, globs.FONT, isinstance(widget, QtW.QPushButton))
+        for widget in main_window.gui_structure.flex_option.frame.children()
+        if isinstance(widget, QtW.QWidget)
+    ]
 
     # test file not found error
     with raises(FileNotFoundError):
         assert globs.path == globs.get_path_for_file(globs.path.joinpath("./ScenarioGUI/gui_classes/gui_structure_classes"), "not_exists.ini")
+
+    config = ConfigParser()
+    config.read(path.parent.joinpath("setup.cfg"))
+    version = config["metadata"]["version"]
+    assert version == globs.VERSION
+
+    load_config(path.joinpath("gui_config_with_version.ini"))
+
+    assert globs.VERSION == "0.1.1.1.1"
 
     close_tests(main_window, qtbot)